Consider adding Pod-Reaper as a Notable Tool
Pod-Reaper is a Chaos Tool for killing pods in kubernetes.  Deploying a Pod Reaper to your kubernetes namespace is a great way to validate your application and micro services can be resilient to chaos kills.  Pod Reaper is designed using environment variables that can be configured to support the chaos you require for testing (once an hour, once a minute, % of pods, etc.).  Tool was authored by Github user murasaki/BrianBerzins.