add Muxy chaos testing tool
Muxy is a proxy for simulating real-world distributed system failures to improve resilience in your applications.

It "mucks" with your system and application context, operating at Layers 4, 5 and 7, allowing you to simulate common failure scenarios from the perspective of an application under test; such as an API or a web application.